services/signal/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
# torch is installed via Dockerfile command for CPU optimization
scikit-learn>=1.3.0
shap>=0.42.0
redis>=5.0.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
pandas_ta>=0.3.14
